requirements.txt
numpy==1.26.4
opencv-python==4.8.1.78
flask==3.1.3
google-genai
python-dotenv
tensorflow==2.13.0
